research/requirements.txt
# Unified dependencies for Pi 4 / Pi 5 (USB camera + NCNN CPU inference)
opencv-python
ncnn
numpy<2
pyzmq
pyyaml
requests
RPi.GPIO
supabase
pyserial
pynmea2
